+++ src/app/App.jsx
@@ -146,6 +146,7 @@ function AppRoutes({ rootElement, publicExplorationRoute, privateRedirect }) {
           <Route path="/escolher-senadores/renovar" element={privateRedirect(BALLOT_ROUTES.senadores)} />
           <Route path={BALLOT_ROUTES.meuPlano} element={publicExplorationRoute(<MeuPlano />)} />
           <Route path={`${BALLOT_ROUTES.continuarPlano}/:token`} element={<ContinuarPlano />} />
+          <Route path={`${BALLOT_ROUTES.continuarPlanoCurto}/:token`} element={<ContinuarPlano />} />
           <Route path="/meu-nossovoto" element={privateRedirect(BALLOT_ROUTES.meuPlano)} />
 
           <Route path="/meu-voto" element={privateRedirect(BALLOT_ROUTES.meuPlano)} />